Implement Stringer. Removed unused files
