--- 02_bubble/SandhyaSMani/02_bubble.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"io"
-	"os"
-)
-
-var out io.Writer = os.Stdout
-
-func main() {
-	fmt.Fprintln(out, bubble([]int{3, 4, 1, 5, 2}))
-}
-
-func bubble(s []int) []int {
-	for i := 1; i < len(s); i++ {
-		for j := 0; j < len(s)-1; j++ {
-			if s[j] > s[j+1] {
-				s[j], s[j+1] = s[j+1], s[j]
-			}
-		}
-	}
-	return s
-}
